backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
cors==1.0.1
httpx==0.24.1
google-genai==0.8.0
